rename physics -> Simulation_plugin
