+++ src/gravity/mod.rs
@@ -8,10 +8,10 @@ use crate::communication::WorldRank;
 use crate::components::Position;
 use crate::components::Velocity;
 use crate::domain::TopLevelIndices;
-use crate::physics::Timestep;
 use crate::prelude::Particles;
 use crate::quadtree::Node;
 use crate::quadtree::*;
+use crate::simulation_plugin::Timestep;
 use crate::units;
 use crate::units::Dimensionless;
 use crate::units::Length;
