+++ src/gravity/plugin.rs
@@ -10,9 +10,9 @@ use crate::communication::Identified;
 use crate::domain::communicate_mass_moments_system;
 use crate::domain::construct_quad_tree_system;
 use crate::named::Named;
-use crate::physics::PhysicsStages;
 use crate::simulation::RaxiomPlugin;
 use crate::simulation::Simulation;
+use crate::simulation_plugin::SimulationStages;
 
 #[derive(Named)]
 pub struct GravityPlugin;
@@ -20,13 +20,13 @@ pub struct GravityPlugin;
 impl RaxiomPlugin for GravityPlugin {
     fn build_everywhere(&self, sim: &mut Simulation) {
         sim.add_parameter_type::<GravityParameters>()
-            .add_system_to_stage(PhysicsStages::Physics, construct_quad_tree_system)
+            .add_system_to_stage(SimulationStages::Physics, construct_quad_tree_system)
             .add_system_to_stage(
-                PhysicsStages::Physics,
+                SimulationStages::Physics,
                 communicate_mass_moments_system.after(construct_quad_tree_system),
             )
             .add_system_to_stage(
-                PhysicsStages::Physics,
+                SimulationStages::Physics,
                 gravity_system.after(communicate_mass_moments_system),
             )
             .add_plugin(
