+++ src/hydrodynamics/mod.rs
@@ -14,11 +14,11 @@ use crate::components::Velocity;
 use crate::domain::extent::Extent;
 use crate::named::Named;
 use crate::performance_parameters::PerformanceParameters;
-use crate::physics::Timestep;
 use crate::prelude::LocalParticle;
 use crate::prelude::Particles;
 use crate::simulation::RaxiomPlugin;
 use crate::simulation::Simulation;
+use crate::simulation_plugin::Timestep;
 use crate::units::helpers::VecQuantity;
 use crate::units::Density;
 use crate::units::Dimension;
