+++ src/parameters.rs
@@ -5,6 +5,6 @@ pub use crate::hydrodynamics::InitialGasEnergy;
 pub use crate::io::input::InputParameters;
 pub use crate::io::output::parameters::*;
 pub use crate::performance_parameters::PerformanceParameters;
-pub use crate::physics::SimulationParameters;
 pub use crate::quadtree::QuadTreeConfig;
+pub use crate::simulation_plugin::SimulationParameters;
 pub use crate::visualization::VisualizationParameters;
