+++ src/prelude.rs
@@ -5,10 +5,10 @@ pub use crate::hydrodynamics::HydrodynamicsPlugin;
 pub use crate::named::*;
 pub use crate::particle::LocalParticle;
 pub use crate::particle::Particles;
-pub use crate::physics::Timestep;
 pub use crate::rand::gen_range;
 pub use crate::simulation::Simulation;
 pub use crate::simulation_builder::SimulationBuilder;
+pub use crate::simulation_plugin::Timestep;
 pub use crate::units;
 pub use crate::units::helpers::Float;
 pub use crate::units::helpers::MVec;
